PB/requirements.txt
django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pillow==9.3.0
geoip2==4.6.0
geopy==2.2.0
pgeocode==0.3.0

